# swagger.yaml
openapi: 3.0.3
info:
  title: EPOS Backoffice RESTful APIs
  description: This is the Backoffice RESTful APIs Swagger page.
  contact:
    email: apis@lists.epos-ip.org
  license:
    name: MIT License
    url: https://epos-ci.brgm.fr/epos/WebApi/raw/master/LICENSE
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/backoffice/v1
    description: Inferred Url
paths:
...
  /person:
    post:
      tags:
        - V3 - Person
      summary: "Create Person instance"
      description: ""
      operationId: "personpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/person'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - Person
      summary: "Update Person"
      description: ""
      operationId: "personputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/person'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /person/{instance_id}:
    get:
      tags:
        - V3 - Person
      summary: "Retrive Person"
      description: ""
      operationId: "personget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/person'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - Person
      summary: "Delete Person"
      description: ""
      operationId: "persondelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/person'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /organization:
    post:
      tags:
        - V3 - Organization
      summary: "Create Organization instance"
      description: ""
      operationId: "organizationpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/organization'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - Organization
      summary: "Update Organization"
      description: ""
      operationId: "organizationputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/organization'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /organization/{instance_id}:
    get:
      tags:
        - V3 - Organization
      summary: "Retrive Organization"
      description: ""
      operationId: "organizationget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/organization'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - Organization
      summary: "Delete Organization"
      description: ""
      operationId: "organizationdelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/organization'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
components:
  schemas:
...
    organization:
      required:
        - UID
        - identifier
      properties:
        UID:
          type: string
          example: 'UID'
          description: 'This property contains the main identifier for the Organization.'
        address:
          $ref: '#/components/schemas/address'
        contactPoint:
          type: array
          items:
            type: string
            example: 'contact Point'
          description: 'This property refers to the Contact Point (i.e. Role) defined for the Organization'
        email:
          type: array
          items:
            type: string
            example: 'email'
          description: 'This property contains the email address of the Organization.'
        identifier:
          type: array
          items:
            $ref: '#/components/schemas/identifier'
          description: 'This property contains an identifier for the Organization (e.g., PIC, ISNI, etc.).'
        legalName:
          type: string
          example: 'legalName'
          description: 'This property contains the official name of the Organization.'
        leiCode:
          type: string
          example: 'leiCode'
          description: 'This property contains the Organization identifier that uniquely identifies a legal entity as defined in ISO 17442.'
        logo:
          type: string
          example: 'logo'
          description: 'This property contains the URL of the Organization logo.'
        memberOf:
          type: array
          items:
            type: string
            example: 'organization'
          description: 'This property refers to an Organization to which this Organization belongs.'
        owns:
          type: array
          items:
            type: string
            example: 'equipment or facility'
          description: 'This property refers to a Facility or Equipment of which it is the owner.'
        telephone:
          type: array
          items:
            type: string
            example: 'telephone'
          description: 'This property contains the telephone number of the Organization.'
        URL:
          type: string
          example: 'URL'
          description: 'This property contains the URL of the Organization website.'
    person:
      required:
        - UID
        - identifier
        - familyName
        - givenName
      properties:
        UID:
          type: string
          example: 'UID'
          description: 'This property contains the main identifier for the Person.'
        address:
          $ref: '#/components/schemas/address'
        affiliation:
          type: array
          items:
            type: string
            example: 'organization'
          description: 'This property refers to the organization to which the person is affiliated.'
        CVURL:
          type: string
          example: 'https://mycv.com/mario.rossi'
          description: 'This property refers to the URL of the Person curriculum vitae.'
        email:
          type: array
          items:
            type: string
            example: 'mario.rossi@gmail.com'
          description: 'This property contains the email address of the Person.'
        familyName:
          type: string
          example: 'Rossi'
          description: 'This property contains the last name of the Person.'
        givenName:
          type: string
          example: 'Mario'
          description: 'This property contains the first name of the Person.'
        identifier:
          type: array
          items:
            $ref: '#/components/schemas/identifier'
          description: 'This property contains an identifier for the Person (e.g., ORCID, ScopusID, etc.).'
        qualification:
          type: string
          example: 'boss'
          description: 'This property contains the specific qualifications of the Person.'
        telephone:
          type: array
          items:
            type: string
            example: '+39 334 86 06 438'
          description: 'This property contains the telephone number of the Person.'
...
    identifier:
      type: object
      required:
        - identifier
        - type
      properties:
        type:
          type: string
          example: 'DDSS-ID'
          description: 'This property refers the identifier scheme referenced by its datatype (e.g., DDSS-ID, DOI, etc.).'
        identifier:
          type: string
          example: 'id'
          description: 'This property contains a string that is an identifier in the context of the identifier scheme referenced by its datatype'
    address:
      type: object
      required:
        - country
        - locality
        - postalCode
        - street
      properties:
        country:
          type: string
          example: 'country'
          description: 'The country. For example, USA. You can also provide the two-letter ISO 3166-1 alpha-2 country code.'
        locality:
          type: string
          example: 'locality'
          description: 'The locality in which the street address is, and which is in the region. For example, Mountain View.'
        postalCode:
          type: string
          example: 'postalCode'
          description: 'The postal code. For example, 94043.'
        street:
          type: string
          example: 'street'
          description: 'The street address. For example, 1600 Amphitheatre Pkwy.'
...
  securitySchemes:
    aai:
      type: apiKey
      in: header
      name: Authorization